2019.05.17 - 최종 팀프로젝트 발표/final_project/workspace/Movies/requirements.txt
awscli==1.16.159
awsebcli==3.15.1
blessed==1.15.0
bootstrap4==0.1.0
botocore==1.12.149
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
Django==2.1.8
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
pathspec==0.5.9
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
virtualenv==16.6.0
wcwidth==0.1.7
websocket-client==0.56.0
